+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Extrapolation parameter in spline transformation
+
 ## [1.2.1] - 2025-03-11
 
 ### Changed